create playlist controller
